+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Enhancement
 
 Refactor
 - move gas skills into separate file
+- abstract fetchGas function
 
 # 0.20 2012/12/17